+++ identity/src/main/java/com/stripe/android/identity/viewmodel/IdentityViewModel.kt
@@ -959,13 +959,13 @@ internal class IdentityViewModel constructor(
                             "sessionID: ${verificationArgs.verificationSessionId} and ephemeralKey: " +
                                 verificationArgs.ephemeralKeySecret
                             ).let { msg ->
-                                _verificationPage.postValue(
-                                    Resource.error(
-                                        msg,
-                                        IllegalStateException(msg, it)
-                                    )
+                            _verificationPage.postValue(
+                                Resource.error(
+                                    msg,
+                                    IllegalStateException(msg, it)
                                 )
-                            }
+                            )
+                        }
                 }
             )
         }