+++ identity/src/main/java/com/stripe/android/identity/viewmodel/IdentityViewModel.kt
@@ -473,7 +473,14 @@ internal class IdentityViewModel constructor(
 
             else -> {
                 Log.e(TAG, "incorrect category: ${legacyOutput.category}")
-                throw IllegalStateException("incorrect targetScanType: ${legacyOutput.category}")
+                isFront = true
+                targetScanType = IdentityScanState.ScanType.DOC_FRONT
+                logError(
+                    IllegalStateException(
+                        "incorrect legacy targetScanType: ${legacyOutput.category}, " +
+                            "upload as DOC_FRONT"
+                    )
+                )
             }
         }
 
@@ -529,7 +536,14 @@ internal class IdentityViewModel constructor(
 
             else -> {
                 Log.e(TAG, "incorrect category: ${modernOutput.category}")
-                throw IllegalStateException("incorrect targetScanType: ${modernOutput.category}")
+                isFront = true
+                targetScanType = IdentityScanState.ScanType.DOC_FRONT
+                logError(
+                    IllegalStateException(
+                        "incorrect modern targetScanType: ${modernOutput.category}, " +
+                            "upload as DOC_FRONT"
+                    )
+                )
             }
         }